Fix conversion from double to int warning. BUG=none TEST=none TBR=jleyba
Review URL: http://codereview.chromium.org/6532077

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@75510 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 33e73aa2d4e9be16871985f3a6433a5bdb6e44c6

+++ chrome/test/webdriver/commands/implicit_wait_command_unittest.cc
@@ -81,7 +81,7 @@ TEST(ImplicitWaitCommandTest, SettingImplicitWaits) {
 
   parameters->SetInteger("ms", 1);
   AssertTimeoutSet(test_session, 1, &command);
-  parameters->SetInteger("ms", 2.5);
+  parameters->SetDouble("ms", 2.5);
   AssertTimeoutSet(test_session, 2, &command);
   parameters->SetInteger("ms", 0);
   AssertTimeoutSet(test_session, 0, &command);
